Fix user profile picture deletion bug

+++ backendcrud.js
@@ -322,17 +322,19 @@ app.delete("/user/:id", (req, res) => {
       if (!user) {
         res.status(404).send("User not found");
       } else {
-        const imagePath = path.join(
-          __dirname,
-          `/public/images/${user.profilePicture}`
-        );
-        fs.unlink(imagePath, (err) => {
-          if (err) {
-            console.log("Error deleting file:", err);
-          } else {
-            console.log("File deleted successfully");
-          }
-        });
+        if (user.profilePicture != "noimage.jpg") {
+          const imagePath = path.join(
+            __dirname,
+            `/public/images/${user.profilePicture}`
+          );
+          fs.unlink(imagePath, (err) => {
+            if (err) {
+              console.log("Error deleting file:", err);
+            } else {
+              console.log("File deleted successfully");
+            }
+          });
+        }
       }
       user
         .destroy()
